use problem.extensions.sciml for separation
