+++ petab/v2/C.py
@@ -259,7 +259,7 @@ MAPPING_FILES = "mapping_files"
 #: Extensions key in the YAML file
 EXTENSIONS = "extensions"
 #: PEtab SciML extension
-SCIML = "sciml"
+EXT_ID_SCIML = "sciml"
 
 
 # MAPPING
