+++ petab/v2/core.py
@@ -54,16 +54,6 @@ from ..v1.yaml import get_path_prefix
 from ..versions import parse_version
 from . import C, get_observable_df
 
-try:
-    from petab_sciml import (
-        ArrayData,
-        ArrayDataStandard,
-        NNModel,
-        NNModelStandard,
-    )
-except ModuleNotFoundError:
-    pass
-
 if TYPE_CHECKING:
     from ..v2.lint import ValidationResultList, ValidationTask
 
@@ -71,7 +61,6 @@ if TYPE_CHECKING:
 __all__ = [
     "Problem",
     "ProblemConfig",
-    "SciMLConfig",
     "Observable",
     "ObservableTable",
     "NoiseDistribution",
@@ -318,6 +307,22 @@ class BaseTable(BaseModel, Generic[T]):
             )
         self.elements.append(other)
         return self
+
+
+# SciML extension classes — imported after BaseTable is defined to avoid
+# circular imports (sciml.py does not import from core.py).
+from .extensions.sciml import (  # noqa: E402
+    HybridizationTable,
+    SciMLConfig,
+    SciMLExt,
+)
+
+
+class ProblemExtensions:
+    """Runtime extension state attached to a :class:`Problem`."""
+
+    def __init__(self, sciml: SciMLExt = None):
+        self.sciml: SciMLExt = sciml or SciMLExt()
 
 
 class Observable(BaseModel):
@@ -1120,77 +1125,6 @@ class ParameterTable(BaseTable[Parameter]):
         return sum(p.estimate for p in self.parameters)
 
 
-class Hybridization(BaseModel):
-    """Assigns PEtab SciML NN inputs and outputs."""
-
-    #: The target ID.
-    target_id: str = Field(alias=C.TARGET_ID)
-    #: The target value.
-    target_value: sp.Basic = Field(alias=C.TARGET_VALUE)
-
-    #: :meta private:
-    model_config = ConfigDict(
-        arbitrary_types_allowed=True,
-        populate_by_name=True,
-        extra="allow",
-        validate_assignment=True,
-    )
-
-    @field_validator("target_value", mode="before")
-    @classmethod
-    def _sympify(cls, v):
-        if v is None or isinstance(v, sp.Basic):
-            return v
-        if isinstance(v, float) and np.isnan(v):
-            return None
-
-        return sympify_petab(v)
-
-
-class HybridizationTable(BaseTable[Hybridization]):
-    """PEtab SciML hybridization table."""
-
-    @property
-    def hybridizations(self) -> list[Hybridization]:
-        """List of hybridizations."""
-        return self.elements
-
-    @classmethod
-    def from_df(cls, df: pd.DataFrame, **kwargs) -> HybridizationTable:
-        """Create a HybridizationTable from a DataFrame."""
-        if df is None:
-            return cls(**kwargs)
-
-        hybridizations = [
-            Hybridization(
-                **row.to_dict(),
-            )
-            for _, row in df.iterrows()
-        ]
-
-        return cls(hybridizations, **kwargs)
-
-    def to_df(self) -> pd.DataFrame:
-        """Convert the HybridizationTable to a DataFrame."""
-        records = self.model_dump(by_alias=True)["elements"]
-
-        return pd.DataFrame(records)
-
-    def __getitem__(self, target_id: str) -> Hybridization:
-        """Get a hybridization by target ID."""
-        for hybridization in self.hybridizations:
-            if hybridization.target_id == target_id:
-                return hybridization
-        raise KeyError(f"Target ID {target_id} not found")
-
-    def get(self, target_id, default=None):
-        """Get a hybridization by target ID or return a default value."""
-        try:
-            return self[target_id]
-        except KeyError:
-            return default
-
-
 class Problem:
     """
     PEtab parameter estimation problem
@@ -1217,16 +1151,18 @@ class Problem:
         measurement_tables: list[MeasurementTable] = None,
         parameter_tables: list[ParameterTable] = None,
         mapping_tables: list[MappingTable] = None,
-        neural_networks: list[NNModel] | None = None,
-        hybridization_tables: list[HybridizationTable] | None = None,
-        array_data_files: list[ArrayData] | None = None,
+        extensions: ProblemExtensions = None,
         config: ProblemConfig = None,
     ):
         from ..v2.lint import default_validation_tasks, sciml_validation_tasks
 
         self.config = config
         self.models: list[Model] = models or []
-        if config and config.extensions and config.extensions[C.SCIML]:
+        if (
+            config
+            and config.extensions
+            and config.extensions.get(C.EXT_ID_SCIML)
+        ):
             self.validation_tasks: list[ValidationTask] = (
                 sciml_validation_tasks.copy()
             )
@@ -1241,11 +1177,7 @@ class Problem:
         self.measurement_tables = measurement_tables or [MeasurementTable()]
         self.mapping_tables = mapping_tables or [MappingTable()]
         self.parameter_tables = parameter_tables or [ParameterTable()]
-        self.neural_networks = neural_networks or []
-        self.hybridization_tables = hybridization_tables or [
-            HybridizationTable()
-        ]
-        self.array_data_files = array_data_files or []
+        self.extensions = extensions or ProblemExtensions()
 
     def __repr__(self):
         return f"<{self.__class__.__name__} id={self.id!r}>"
@@ -1418,10 +1350,10 @@ class Problem:
             else None
         )
 
-        neural_networks = None
-        hybridization_tables = None
-        array_data_files = None
-        if config.extensions and config.extensions[C.SCIML]:
+        extensions = ProblemExtensions()
+        if config.extensions and config.extensions.get(C.EXT_ID_SCIML):
+            from petab_sciml import ArrayDataStandard, NNModel, NNModelStandard
+
             # Neural network classes are constructed via pytorch for now to get
             # the proper inputs
             neural_networks = [
@@ -1435,19 +1367,27 @@ class Problem:
                     nn_model_id=nn_id,
                 )
                 for nn_id, nn_config in (
-                    config.extensions[C.SCIML].neural_networks or {}
+                    config.extensions[C.EXT_ID_SCIML].neural_networks or {}
                 ).items()
             ]
 
             hybridization_tables = [
                 HybridizationTable.from_tsv(f, base_path)
-                for f in config.extensions[C.SCIML].hybridization_files
+                for f in config.extensions[C.EXT_ID_SCIML].hybridization_files
             ]
 
             array_data_files = [
                 ArrayDataStandard.load_data(_generate_path(f, base_path))
-                for f in config.extensions[C.SCIML].array_files
+                for f in config.extensions[C.EXT_ID_SCIML].array_files
             ]
+
+            extensions = ProblemExtensions(
+                sciml=SciMLExt(
+                    neural_networks=neural_networks,
+                    hybridization_tables=hybridization_tables,
+                    array_data_files=array_data_files,
+                )
+            )
 
         return Problem(
             config=config,
@@ -1458,9 +1398,7 @@ class Problem:
             measurement_tables=measurement_tables,
             parameter_tables=parameter_tables,
             mapping_tables=mapping_tables,
-            neural_networks=neural_networks,
-            hybridization_tables=hybridization_tables,
-            array_data_files=array_data_files,
+            extensions=extensions,
         )
 
     @staticmethod
@@ -1767,34 +1705,6 @@ class Problem:
             self.config = ProblemConfig(format_version="2.0.0")
         self.config.id = value
 
-    @property
-    def hybridizations(self) -> list[Hybridization]:
-        """
-        List of hybridizations in the hybridization table(s).
-        Note that hybridizations are specific to PEtab SciML problems.
-        """
-        return list(
-            chain.from_iterable(
-                ht.hybridizations for ht in self.hybridization_tables
-            )
-        )
-
-    @property
-    def hybridization_df(self) -> pd.DataFrame | None:
-        """
-        Combined SciML hybridization tables as DataFrame.
-        Note that hybridizations are specific to PEtab SciML problems.
-        """
-        return (
-            HybridizationTable(hybridizations).to_df()
-            if (hybridizations := self.hybridizations)
-            else None
-        )
-
-    @hybridization_df.setter
-    def hybridization_df(self, value: pd.DataFrame):
-        self.hybridization_tables = [HybridizationTable.from_df(value)]
-
     def get_optimization_parameters(self) -> list[str]:
         """
         Get the list of optimization parameter IDs from parameter table.
@@ -2099,7 +2009,7 @@ class Problem:
 
         validation_results = ValidationResultList()
 
-        supported_extensions = {C.SCIML}
+        supported_extensions = {C.EXT_ID_SCIML}
         if (
             self.config
             and self.config.extensions
@@ -2411,73 +2321,6 @@ ExperimentPeriod(time=2.0, condition_ids=['condition2a', 'condition2b'])])
             Experiment(id=id_, periods=periods)
         )
 
-    def add_hybridization(self, target_id: str, target_value: str):
-        """Add a SciML hybridization table entry to the problem.
-
-        If there is more than one hybridization table, the hybridization
-        is added to the last table. Note that hybridizations are specific
-        to PEtab SciML problems.
-
-        Arguments:
-            target_id: The ID of the target entity in the PEtab problem
-                or neural network model
-            target_value: The value that is assigned to the target id.
-        """
-        if not self.hybridization_tables:
-            self.hybridization_tables.append(HybridizationTable())
-        self.hybridization_tables[-1].hybridizations.append(
-            Hybridization(target_id=target_id, target_value=target_value)
-        )
-
-    def add_neural_network_from_dict(self, model_id: str, nn_dict: dict):
-        """
-        Add a SciML neural net from a dictionary (or PEtab SciML problems).
-        """
-        # from petab_sciml import NNModel
-        nn_model = NNModel.model_validate(nn_dict)
-        nn_model.nn_model_id = model_id
-        self.neural_networks.append(nn_model)
-
-    def add_neural_network_from_yaml(
-        self,
-        model_id: str,
-        file_path: str | Path,
-        base_path: str | Path | None = None,
-    ):
-        """
-        Add a SciML neural net from a yaml file (for PEtab SciML problems).
-        """
-        # from petab_sciml import NNModelStandard
-        self.neural_networks.append(
-            NNModelStandard.load_data(
-                _generate_path(
-                    file_path=file_path,
-                    base_path=base_path,
-                ),
-                nn_model_id=model_id,
-            )
-        )
-
-    def add_array_data_from_dict(self, array_data: dict):
-        """
-        Add SciML array data from a dictionary (for PEtab SciML problems).
-        """
-        # from petab_sciml import ArrayData
-        self.array_data_files.append(ArrayData.model_validate(array_data))
-
-    def add_array_data_from_hdf5(
-        self,
-        file_path: str | Path,
-        base_path: str | Path | None = None,
-    ):
-        """
-        Add SciML array data from an hdf5 file (for PEtab SciML problems).
-        """
-        # from petab_sciml import ArrayDataStandard
-        self.array_data_files.append(
-            ArrayDataStandard.load_data(_generate_path(file_path, base_path))
-        )
-
     def __iadd__(self, other):
         """Add Observable, Parameter, Measurement, Condition, or Experiment"""
         from .core import (
@@ -2682,42 +2525,11 @@ class ModelFile(BaseModel):
     )
 
 
-class NeuralNetConfig(BaseModel):
-    """A neural net in the PEtab SciML problem configuration."""
-
-    location: AnyUrl | Path
-    pre_initialization: bool
-    format: str
-
-    model_config = ConfigDict(
-        validate_assignment=True,
-    )
-
-
 class ExtensionConfig(BaseModel):
     """The configuration of a PEtab extension."""
 
     version: str
     config: dict
-
-
-class SciMLConfig(BaseModel):
-    """The extended configuration of a PEtab SciML problem."""
-
-    #: The PEtab SciML format version.
-    version: str = "0.1.0"
-    #: The paths to the array data files.
-    # Absolute or relative to `base_path`.
-    array_files: list[AnyUrl | Path] = []
-    #: The paths to the hybridization tables.
-    # Absolute or relative to `base_path`.
-    hybridization_files: list[AnyUrl | Path] = []
-    #: The neural network IDs and info.
-    neural_networks: dict[str, NeuralNetConfig] | None = {}
-
-    model_config = ConfigDict(
-        validate_assignment=True,
-    )
 
 
 class ProblemConfig(BaseModel):
@@ -2777,7 +2589,7 @@ class ProblemConfig(BaseModel):
         if isinstance(v, dict):
             parsed_extensions = {}
             for ext_name, ext_config in v.items():
-                if ext_name == C.SCIML:
+                if ext_name == C.EXT_ID_SCIML:
                     # Convert sciml extension to SciMLConfig
                     parsed_extensions[ext_name] = SciMLConfig(**ext_config)
                 else:
@@ -2830,7 +2642,7 @@ class ProblemConfig(BaseModel):
             del data["id"]
 
         for ext_id, d_ext in data[C.EXTENSIONS].items():
-            if ext_id == C.SCIML:
+            if ext_id == C.EXT_ID_SCIML:
                 # convert Paths to strings
                 for key in ("array_files", "hybridization_files"):
                     d_ext[key] = list(map(str, d_ext[key]))
